+++ src/Futhark/CodeGen/Backends/GenericC.hs
@@ -84,6 +84,7 @@ import Control.Monad.Identity
 import Control.Monad.Reader
 import Control.Monad.State
 import Data.Bifunctor (first)
+import Data.Char (isAlpha, isAlphaNum)
 import qualified Data.DList as DL
 import Data.List (unzip4)
 import Data.Loc
@@ -99,6 +100,7 @@ import Futhark.CodeGen.RTS.C (cacheH, errorsH, halfH, lockH, timingH, utilH)
 import Futhark.IR.Prop (isBuiltInFunction)
 import qualified Futhark.Manifest as Manifest
 import Futhark.MonadFreshNames
+import Futhark.Util (zEncodeString)
 import Futhark.Util.Pretty (prettyText)
 import qualified Language.C.Quote.OpenCL as C
 import qualified Language.C.Syntax as C
@@ -1335,6 +1337,18 @@ prepareEntryOutputs = collect' . zipWithM prepare [(0 :: Int) ..]
             [C.cstm|$exp:dest->shape[$int:i] = $id:d;|]
       stms $ zipWith maybeCopyDim shape [0 .. rank - 1]
 
+isValidCName :: Name -> Bool
+isValidCName = check . nameToString
+  where
+    check [] = True -- academic
+    check (c : cs) = isAlpha c && all constituent cs
+    constituent c = isAlphaNum c || c == '_'
+
+entryName :: Name -> String
+entryName v
+  | isValidCName v = "entry_" <> nameToString v
+  | otherwise = "entry_" <> zEncodeString (nameToString v)
+
 onEntryPoint ::
   [C.BlockItem] ->
   Name ->
@@ -1349,7 +1363,7 @@ onEntryPoint get_consts fname (Function (Just ename) outputs inputs _ results ar
   outputdecls <- collect $ mapM_ stubParam outputs
   decl_mem <- declAllocatedMem
 
-  entry_point_function_name <- publicName $ "entry_" ++ nameToString ename
+  entry_point_function_name <- publicName $ entryName ename
 
   (inputs', unpack_entry_inputs) <- prepareEntryInputs $ map snd args
   let (entry_point_input_params, entry_point_input_checks) = unzip inputs'