+++ src/Language/Futhark/TypeChecker.hs
@@ -624,13 +624,6 @@ entryPoint params orig_ret_te (RetType ret orig_ret) =
     onRetType te t =
       ([], EntryType t te)
 
-entryPointNameIsAcceptable :: Name -> Bool
-entryPointNameIsAcceptable = check . nameToString
-  where
-    check [] = True -- academic
-    check (c : cs) = isAlpha c && all constituent cs
-    constituent c = isAlphaNum c || c == '_'
-
 checkValBind :: ValBindBase NoInfo Name -> TypeM (Env, ValBind)
 checkValBind (ValBind entry fname maybe_tdecl NoInfo tparams params body doc attrs loc) = do
   top_level <- atTopLevel
@@ -646,8 +639,6 @@ checkValBind (ValBind entry fname maybe_tdecl NoInfo tparams params body doc att
 
   case entry' of
     Just _
-      | not $ entryPointNameIsAcceptable fname ->
-          typeError loc mempty "Entry point names must start with a letter and contain only letters, digits, and underscores."
       | any isTypeParam tparams' ->
           typeError loc mempty "Entry point functions may not be polymorphic."
       | not (all patternOrderZero params')